src/backend/requirements.txt
Django==4.1.2
channels==3.0.5
channels-redis==4.0.0
daphne==3.0.2
Pillow==9.2.0
django-oauth-toolkit>=1.3
hvac==0.11.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.1
Deprecated==1.2.13
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
psycopg2
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
redis==4.3.4
rich==12.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
wrapt==1.14.1
zope.interface==5.5.0
djangorestframework_simplejwt
